-Все контроллеры работают, кроме создания пользователя

+++ Cogni/Abstractions/Repositories/IUserRepository.cs
@@ -6,7 +6,7 @@ namespace Cogni.Abstractions.Repositories
     {
         Task<UserModel> Get(string login, string passwordhash);//вход
         //Task<UserModel> Get(long id); //для профиля
-        Task<long> Create(Customuser user);
+        Task<int> Create(Customuser user);
         Task<bool> CheckUser(string login);//проверка существования пользователя с таким логином
     }
 }
